Update Opal_Leganyn.lua
Copying the format for one of the Holy Symbol Of Tunare quest turn-ins, this is an attempt to create a really simple little turn-in to upgrade the disappointing falchion you get from General Jyleel: https://wiki.project1999.com/The_Falchion
I've altered her dialogue text to provide accurate instructions and commentary, but the trigger phrase might be malformed (I left the trigger phrase alone for now). Will try to alter General Jyleel's files to provide directions to take the falchion to this NPC after he awards it.